requirements.txt
aiogram==2.25.1
asyncpg==0.29.0
APScheduler==3.10.4
python-dotenv==1.0.1
